refactor: split codegen options
